requirements.txt
flask
opencv-python-headless
numpy
torch
torchvision
facenet-pytorch
faiss-cpu
